Move settings.py into RimSort.models
